move rail logic to delayed and updates to some css to reduce CLS
